PLANET-7849: Refactor header.js file
- Rename the accessibleNavMenu.js file
